Удаляет layout
