+++ src/components/app/app.tsx
@@ -5,15 +5,11 @@ import {Redirect, Route, Router, Switch} from 'react-router-dom';
 import {User} from "../../models/User";
 import history from "../../history";
 import {getUserInfo} from "../../reducer/user/selectors";
-import {withLayout} from '../../hocs/with-layout/with-layout';
 import MainPage from '../pages/main/main';
 import FilmPage from '../pages/film/film';
 import {Login} from '../pages/login/login';
 import MyList from "../pages/my-list/my-list";
-import LayoutMainPage from "../shared/layout-main-page/layout-main-page";
 
-
-const MainPageWrapped = withLayout(MainPage, LayoutMainPage);
 
 interface PrivateRouteProps {
   component: React.ComponentType;
@@ -48,7 +44,7 @@ export const App: React.FC<Props> = (props) => {
     <Router history={history}>
       <Switch>
         <Route path="/" exact
-          render={(mainPageProps): React.ReactNode => <MainPageWrapped {...mainPageProps} layoutProps={{user}}/>}
+          render={(mainPageProps): React.ReactNode => <MainPage {...mainPageProps} user={{user}}/>}
         />
         <Route path="/films/:id" exact component={FilmPage}/>
         <PrivateRoute path="/mylist" component={MyList} is={isLogin} redirectTo={`/login`} data={{user}}/>
